remove superfluous GUI class. Now there is just one GUI class : ilcleanUpsessionsConfigGUI it has all the functionality of the previous MainGui
